Don't verify existence of autorenew poll messages
There are several situations in which it is a valid system state for
Keys to PollMessages on Domains to point to entities that don't exist,
chief amongst them when a domain has been soft-deleted.  We also can't
simply null this out when soft-deleting domains, because there may be
an un-acked recurrence meaning it can't be deleted.

Given this, it doesn't make sense to check for entity existence here,
as it is not a system invariant and cannot easily be made to be so.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=122300368