Reorganize DomainTransferRequestFlow helper method structure
This attempts to improve the organization of the helper methods in
DomainTransferRequestFlow created by the flattening in []

The primary changes are:

 - new createTransferServerApproveEntities() method that now makes
   all the server approve entities in one centralized place

 - new createPendingTransferData() method that takes the server
   approve entities and hides the slightly hacky code that parses out
   the right ones to stuff into the TransferData

This seems like an overall simpler structure to me.  With this change,
run() is 20 lines shorter and the flow overall is 40 lines shorter (not
counting a big blob of javadoc I added).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134315295