Reduce Datastore load of delete prober data mapreduce
This eliminates the transactional load of ForeignKeyIndexes and
EppResourceIndexes, the latter of which was problematic because it is parented
on the EppResourceIndexBucket entity group, and can cause concurrent
modification exceptions on live code paths. By removing the transactional load
and only touching that entity group on the delete, the number of potential
concurrent modification exceptions is significantly reduced.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135706974