Remove LINKED from the disallowed statuses for contact/host delete
LINKED is supposed to be a virtual status that gets added/removed
when needed. It's mistakenly been persisted to datastore on many
resources, but the persisted value is meaningless and may not
represent reality at all. There is no reason to check for LINKED
status before kicking off a delete, since the smoke test and
[] will catch all actual linked objects, and the LINKED
status can be causing false positives for objects that are no
longer LINKED.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=145422457