Fix connection error in nomulus tool caused by YAML configs
This is a temporary work-around that fixes the tool for all of our
environments. Next up I'm working on a generalizable way to get this
working by adding some kind of configuration between environment name
and App Engine project ID. The current configuration system doesn't
quite work for that because it's all based on a separate config per
environment, whereas the tool needs to be able to access all
environments. Either we bundle all configs that currently go into
WEB-INF/ with nomulus and have it select based on the -e flag, or we
make it a separate configuration.

TESTED=I built and ran locally and was able to successfully run
commands against alpha, production, and sandbox.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=146481850